# storage: disable key access assertions in non-race builds