@petermattis 2017-05-26
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 3 files reviewed at latest revision, all discussions resolved, some commit checks pending.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/16166#-:-Kl45SNNzIACHdfzgQS9:bnfp4nl)*
<!-- Sent from Reviewable.io -->